Adding Transitions

+++ Core/Pashmak/VideoFile.cpp
@@ -23,7 +23,7 @@ bool VideoFile::GetNextFrame(std::shared_ptr<Picture>& frame)
 
 unsigned int VideoFile::GetLength()
 {
-	return myVideoCapture->get(CV_CAP_PROP_FRAME_COUNT);
+	return static_cast<int>(myVideoCapture->get(CV_CAP_PROP_FRAME_COUNT));
 }
 
 void VideoFile::GotoFrame(unsigned int frame)
